requirements.txt
# Data Manipulation
pandas
numpy

# Data Generation
faker
# azure
azure-storage-blob

# Misc
python-dotenv
